+++ src/main/java/persistence/Application.java
@@ -6,9 +6,9 @@ import jdbc.JdbcTemplate;
 import jdbc.RowMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import persistence.sql.ddl.H2Dialect;
+import persistence.sql.H2Dialect;
 import persistence.sql.ddl.query.CreateQueryBuilder;
-import persistence.sql.ddl.Person;
+import persistence.sql.Person;
 import persistence.sql.ddl.query.DropQueryBuilder;
 
 import java.sql.ResultSet;
