+++ src/main/java/persistence/sql/Person.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql;
 
 import jakarta.persistence.*;
