+++ src/main/java/persistence/sql/Queryable.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql;
 
 public interface Queryable {
     void apply(StringBuilder query, Dialect dialect);
