+++ src/main/java/persistence/sql/ddl/query/AutoKeyGenerationStrategy.java
@@ -2,7 +2,7 @@ package persistence.sql.ddl.query;
 
 import jakarta.persistence.GenerationType;
 import persistence.sql.ddl.PrimaryKeyGenerationStrategy;
-import persistence.sql.ddl.definition.TableId;
+import persistence.sql.definition.TableId;
 
 public class AutoKeyGenerationStrategy implements PrimaryKeyGenerationStrategy {
