+++ src/main/java/persistence/sql/definition/TableDefinition.java
@@ -1,11 +1,11 @@
-package persistence.sql.ddl.definition;
+package persistence.sql.definition;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
 import org.jetbrains.annotations.NotNull;
-import persistence.sql.ddl.Queryable;
+import persistence.sql.Queryable;
 
 import java.lang.reflect.Field;
 import java.util.Arrays;
