+++ src/main/java/persistence/sql/dml/query/DmlQueryBuilder.java
@@ -1,0 +1,5 @@
+package persistence.sql.dml.query;
+
+public interface DmlQueryBuilder {
+    String build(Object entity);
+}
