+++ src/test/java/persistence/sql/ddl/definition/ColumnDefinitionTest.java
@@ -5,7 +5,8 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import persistence.sql.ddl.SqlType;
+import persistence.sql.SqlType;
+import persistence.sql.definition.ColumnDefinition;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
