+++ src/test/java/persistence/sql/ddl/definition/TableDefinitionTest.java
@@ -3,6 +3,7 @@ package persistence.sql.ddl.definition;
 import jakarta.persistence.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import persistence.sql.definition.TableDefinition;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
