+++ src/test/java/persistence/sql/ddl/query/CreateQueryBuilderTest.java
@@ -2,9 +2,9 @@ package persistence.sql.ddl.query;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import persistence.sql.ddl.Dialect;
-import persistence.sql.ddl.H2Dialect;
-import persistence.sql.ddl.Person;
+import persistence.sql.Dialect;
+import persistence.sql.H2Dialect;
+import persistence.sql.Person;
 import persistence.sql.ddl.fixtures.TestEntityWithAutoIdStrategy;
 import persistence.sql.ddl.fixtures.TestEntityWithIdentityIdStrategy;
 import persistence.sql.ddl.fixtures.TestEntityWithNullableColumns;
