+++ src/test/java/persistence/sql/ddl/query/DropQueryBuilderTest.java
@@ -2,7 +2,7 @@ package persistence.sql.ddl.query;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import persistence.sql.ddl.Person;
+import persistence.sql.Person;
 
 import static org.assertj.core.api.Assertions.assertThat;
